requirements.txt
python-telegram-bot
yadisk
python-dotenv
